+++ platform/nxp/lpc43xx/sdmmc_base.c
@@ -174,7 +174,7 @@ static void configPins(struct SdmmcBase *interface,
   };
   bool wide = true;
 
-  for (uint8_t index = 0; index < ARRAY_SIZE(pinArray); ++index)
+  for (unsigned int index = 0; index < ARRAY_SIZE(pinArray); ++index)
   {
     if (!pinArray[index])
     {
